Organização do projeto, separação dos modulos  do express implementos em app.js.
