+++ routes/produtosRoute.js
@@ -1,9 +1,11 @@
+'use strict';
 
-var express = require('express');
-var produtosRouter = express.Router();
-produtosRouter.route('')
-    .get(function (req,res) {
-         res.send("Exibindo a página principal...");
-     });
-
-module.exports = produtosRouter; 
+module.exports = app => {
+    var express = require('express');
+    var produtosRouter = express.Router();
+    produtosRouter.route('')
+        .get(function (req, res) {
+            res.send("Exibindo a página principal...");
+        });
+    app.use('/produtos', produtosRouter)
+}
